requirements.txt
Flask
flask-cors
flask-login
pycryptodome
sqlalchemy
werkzeug
python-dotenv
gunicorn
